+++ src/ResourceManager/HDInsight/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Added support for Data Disks property in cluster creation
+    - Added parameter 'WorkerNodeDataDisksGroups' to the New-AzureHDInsightCluster cmdlet
 
 ## Version 3.3.1